+++ src/Trans.js
@@ -277,6 +277,7 @@ export function Trans({
   count,
   parent,
   i18nKey,
+  context,
   tOptions = {},
   values,
   defaults,
@@ -296,6 +297,8 @@ export function Trans({
   }
 
   const t = tFromProps || i18n.t.bind(i18n) || ((k) => k);
+
+  if (context) tOptions.context = context;
 
   const reactI18nextOptions = { ...getDefaults(), ...(i18n.options && i18n.options.react) };